bit to 1

+++ bit_manipulation/3-set_bit.c
@@ -8,8 +8,8 @@
 */
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	int i = 0;
-	int sum = 1;
+	unsigned int i = 0;
+	unsigned long int sum = 1;
 
 	if (index > 64)
 		return (-1);
